// react-module/src/openapi/models/atomic-operation-extended-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AtomicOperationExtendedDto
 */
export interface AtomicOperationExtendedDto {
    /**
     * 
     * @type {string}
     * @memberof AtomicOperationExtendedDto
     */
    operationResult?: AtomicOperationExtendedDtoOperationResultEnum;
    /**
     * 
     * @type {number}
     * @memberof AtomicOperationExtendedDto
     */
    id?: any;
    /**
     * 
     * @type {OperationTypeExtendedDto}
     * @memberof AtomicOperationExtendedDto
     */
    operationType?: any;
    /**
     * 
     * @type {number}
     * @memberof AtomicOperationExtendedDto
     */
    priority?: any;
    /**
     * 
     * @type {Array&lt;AtomicProductExtendedDto&gt;}
     * @memberof AtomicOperationExtendedDto
     */
    inputProducts?: any;
    /**
     * 
     * @type {Array&lt;AtomicProductExtendedDto&gt;}
     * @memberof AtomicOperationExtendedDto
     */
    outputProducts?: any;
    /**
     * 
     * @type {ProductionOrderExtendedDto}
     * @memberof AtomicOperationExtendedDto
     */
    productionOrder?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicOperationExtendedDto
     */
    groupingOption?: AtomicOperationExtendedDtoGroupingOptionEnum;
}

/**
    * @export
    * @enum {string}
    */
export enum AtomicOperationExtendedDtoOperationResultEnum {
    SUCCESS = 'SUCCESS',
    BROKEN = 'BROKEN',
    ASSIGNED = 'ASSIGNED',
    NOTASSIGNED = 'NOT_ASSIGNED',
    ISRUNNING = 'IS_RUNNING',
    COMPLETED = 'COMPLETED',
    POSTPONED = 'POSTPONED'
}
/**
    * @export
    * @enum {string}
    */
export enum AtomicOperationExtendedDtoGroupingOptionEnum {
    EACHOPERATIONTOSEPARATECYCLE = 'EACH_OPERATION_TO_SEPARATE_CYCLE',
    ALLOPERATIONTOONECYCLE = 'ALL_OPERATION_TO_ONE_CYCLE'
}