// react-module/src/openapi/models/default-input-product-data.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface DefaultInputProductData
 */
export interface DefaultInputProductData {
    /**
     * 
     * @type {string}
     * @memberof DefaultInputProductData
     */
    name?: any;
    /**
     * 
     * @type {Dims}
     * @memberof DefaultInputProductData
     */
    dims?: any;
}