// react-module/src/openapi/models/layer-shape-basic-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface LayerShapeBasicDto
 */
export interface LayerShapeBasicDto {
    /**
     * 
     * @type {number}
     * @memberof LayerShapeBasicDto
     */
    id?: any;
    /**
     * 
     * @type {RectangleShapeBasicDto}
     * @memberof LayerShapeBasicDto
     */
    rectangleShape?: any;
    /**
     * 
     * @type {string}
     * @memberof LayerShapeBasicDto
     */
    layerShapeType?: LayerShapeBasicDtoLayerShapeTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof LayerShapeBasicDto
     */
    thickness?: any;
}

/**
    * @export
    * @enum {string}
    */
export enum LayerShapeBasicDtoLayerShapeTypeEnum {
    RECTANGLE = 'RECTANGLE',
    CIRCLE = 'CIRCLE',
    CUSTOMSHAPE = 'CUSTOM_SHAPE'
}